backend/requirements.txt
fastapi==0.109.2
uvicorn==0.22.0
motor==3.1.1
pymongo==4.3.3
python-jose[cryptography]==3.3.0  # Specifically use python-jose with cryptography
python-multipart==0.0.6
bcrypt==4.0.1
pydantic==2.6.1
azure-storage-blob==12.16.0
python-dotenv==1.0.0
azure-storage-blob==12.16.0
web3==6.15.0
azure-identity